Removed marker decorators from plugin manager and moved to TabTemplate
