+++ PyProma_GUI/PyProma_dir_view/plugins/todo_plugin.py
@@ -13,14 +13,13 @@ import tkinter as tk
 import tkinter.ttk as ttk
 from typing import TYPE_CHECKING
 
-from PyProma_common.PyProma_templates import tab_template
-from PyProma_dir_view.plugins.plugin_manager import PyFileMethod, RefreshMethod
+from PyProma_common.PyProma_templates.tab_template import TabTemplate
 
 if TYPE_CHECKING:
     from PyProma_dir_view.plugins.plugin_manager import PluginManager
 
 
-class TodoTab(tab_template.TabTemplate):
+class TodoTab(TabTemplate):
     NAME = "ToDo"
 
     def __init__(self, master: tk.Tk, main: "PluginManager"):
@@ -29,11 +28,11 @@ class TodoTab(tab_template.TabTemplate):
         self.todo_tree.heading("#0", text="ToDo", anchor=tk.CENTER)
         self.todo_tree.pack(fill=tk.BOTH, expand=True)
 
-    @RefreshMethod
+    @TabTemplate.RefreshMethod
     def refresh(self):
         self.todo_tree.delete(*self.todo_tree.get_children())
 
-    @PyFileMethod
+    @TabTemplate.PyFileMethod
     def find_todo(self, filename: str):
         """this func finds todos and add node to todo_tree.
         this finds "# TODO", "# BUG", "# FIXME", "# HACK".
